+++ src/reducers/appReducer.js
@@ -40,7 +40,6 @@ const initialState = {
     isSmuMode: false,
     deviceRunning: true,
     rttRunning: false,
-    advancedMode: false,
     capabilities: {},
     samplingRunning: false,
     isSaveChoiceDialogVisible: false,
@@ -54,11 +53,9 @@ const RTT_CALLED_START = 'RTT_CALLED_START';
 const SAMPLING_STARTED = 'SAMPLING_STARTED';
 const SAMPLING_STOPPED = 'SAMPLING_STOPPED';
 const SET_POWER_MODE = 'SET_POWER_MODE';
-const TOGGLE_ADVANCED_MODE = 'TOGGLE_ADVANCED_MODE';
 const TOGGLE_SAVE_CHOICE_DIALOG = 'TOGGLE_SAVE_CHOICE_DIALOG';
 const TOGGLE_EXPORT_DIALOG = 'TOGGLE_EXPORT_DIALOG';
 
-export const toggleAdvancedModeAction = () => ({ type: TOGGLE_ADVANCED_MODE });
 export const samplingStartAction = () => ({ type: SAMPLING_STARTED });
 export const samplingStoppedAction = () => ({ type: SAMPLING_STOPPED });
 
@@ -102,10 +99,6 @@ export default (state = initialState, { type, ...action }) => {
         case RTT_CALLED_START: return {
             ...state,
             rttRunning: true,
-        };
-        case TOGGLE_ADVANCED_MODE: return {
-            ...state,
-            advancedMode: !state.advancedMode,
         };
         case TOGGLE_SAVE_CHOICE_DIALOG: return {
             ...state,